Refactor SearchResults.
