+++ components/webui/imports/ui/SearchView/SearchResults/SearchResultsHeader.scss
