+++ components/webui/imports/ui/SearchView/SearchResults/SearchResultsTable.scss
