+++ components/webui/imports/ui/SearchView/SearchView.jsx
@@ -28,8 +28,8 @@ import {unquoteString} from "/imports/utils/misc";
 
 import {LOCAL_STORAGE_KEYS} from "../constants";
 import SearchControls from "./SearchControls/SearchControls.jsx";
-import SearchResults, {VISIBLE_RESULTS_LIMIT_INITIAL} from "./SearchResults.jsx";
-import {computeTimelineConfig} from "./SearchResultsTimeline.jsx";
+import SearchResults, {VISIBLE_RESULTS_LIMIT_INITIAL} from "./SearchResults/SearchResults.jsx";
+import {computeTimelineConfig} from "./SearchResults/SearchResultsTimeline.jsx";
 
 
 // for pseudo progress bar
